update submodules, fix
