+++ src/lib/imports.ts
@@ -4,6 +4,9 @@ import dynamic from 'next/dynamic';
 import * as GQLExamples from '~/docs/fuel-graphql-docs/examples';
 import * as FuelExamples from '~/docs/fuels-wallet/packages/docs/examples';
 
+import { COMPONENTS as GQL_COMPONENTS } from '../component-exports/graphql';
+import { COMPONENTS as WALLET_COMPONENTS } from '../component-exports/wallet';
+
 import type { DocType } from './types';
 
 function loadComponent(imp: any, name?: string) {
@@ -16,44 +19,41 @@ export function getComponents(doc: DocType) {
   const components: any = {};
   const slug = doc.docsConfig.slug || '';
 
-  if (['wallet/', 'guides/'].some((s) => slug.includes(s))) {
+  function addComponents(list: ComponentsList) {
+    Object.keys(list).forEach((page) => {
+      if (doc.slug.includes(page)) {
+        list[page].forEach((comp: any) => {
+          if (comp.import) {
+            components[comp.name] = comp.import;
+          } else if (comp.imports) {
+            components[comp.name] = comp.imports;
+          }
+        });
+      }
+    });
+  }
+
+  if (['guides/'].some((s) => slug.startsWith(s))) {
+    components.CodeImport = loadComponent(
+      import('~/src/components/CodeImport'),
+      'CodeImport'
+    );
+  } else if (['docs/wallet/'].some((s) => slug.startsWith(s))) {
     components.CodeImport = loadComponent(
       import('~/src/components/CodeImport'),
       'CodeImport'
     );
 
-    // load the components used in the wallet docs
-    components.Demo = loadComponent(
-      import('~/docs/fuels-wallet/packages/docs/src/components/Demo'),
-      'Demo'
-    );
-
-    components.Player = loadComponent(
-      import('~/docs/fuels-wallet/packages/docs/src/components/Player'),
-      'Player'
-    );
-
-    components.InstallSection = loadComponent(
-      import('~/docs/fuels-wallet/packages/docs/src/components/InstallSection'),
-      'InstallSection'
-    );
-
-    components.ConnectionAlert = loadComponent(
-      import(
-        '~/docs/fuels-wallet/packages/docs/src/components/ConnectionAlert'
-      ),
-      'ConnectionAlert'
-    );
+    addComponents(WALLET_COMPONENTS);
 
     components.Examples = FuelExamples;
-  }
-
-  if (['graphql/', 'recipes/'].some((s) => slug.includes(s))) {
+  } else if (['docs/graphql/'].some((s) => slug.startsWith(s))) {
     components.GQLExamples = GQLExamples;
     components.CodeExamples = loadComponent(
       import('~/src/components/GraphqlCodeExample'),
       'GraphQLCodeExample'
     );
+    addComponents(GQL_COMPONENTS);
   }
 
   return components;
